separate the game from the engine
