+++ src/com/sulfrix/shroomrun/Main.java
@@ -4,6 +4,7 @@ import processing.core.PApplet;
 
 public class Main {
     public static void main(String[] args) {
-        PApplet.main("com.sulfrix.shroomrun.ShroomRun");
+        PApplet.runSketch(new String[] {"ShroomRun"}, new ShroomRun());
+        //PApplet.main("com.sulfrix.shroomrun.ShroomRun");
     }
 }
