+++ src/com/sulfrix/shroomrun/entities/PhysicsTest.java
@@ -1,7 +1,7 @@
 package com.sulfrix.shroomrun.entities;
 
-import com.sulfrix.shroomrun.Entity;
-import com.sulfrix.shroomrun.lib.BoundingBox;
+import com.sulfrix.sulfur.entity.PhysicsEntity;
+import com.sulfrix.sulfur.lib.BoundingBox;
 import processing.core.PGraphics;
 import processing.core.PVector;
