+++ src/com/sulfrix/shroomrun/entities/Tile.java
@@ -1,8 +1,8 @@
 package com.sulfrix.shroomrun.entities;
 
-import com.sulfrix.shroomrun.Entity;
-import com.sulfrix.shroomrun.lib.GlobalManagers.AssetCache;
-import com.sulfrix.shroomrun.lib.BoundingBox;
+import com.sulfrix.sulfur.entity.Entity;
+import com.sulfrix.sulfur.lib.GlobalManagers.AssetCache;
+import com.sulfrix.sulfur.lib.BoundingBox;
 import processing.core.PConstants;
 import processing.core.PGraphics;
 import processing.core.PImage;
