+++ src/com/sulfrix/shroomrun/scenarios/MenuScenario.java
@@ -1,7 +1,7 @@
 package com.sulfrix.shroomrun.scenarios;
 
-import com.sulfrix.shroomrun.Scenario;
-import com.sulfrix.shroomrun.entities.Camera;
+import com.sulfrix.sulfur.Scenario;
+import com.sulfrix.sulfur.entity.Camera;
 import com.sulfrix.shroomrun.entities.ui.TextEntity;
 import processing.core.PGraphics;
 import processing.core.PVector;
