+++ src/com/sulfrix/shroomrun/scenarios/ParallaxTestScenario.java
@@ -1,7 +1,7 @@
 package com.sulfrix.shroomrun.scenarios;
 
-import com.sulfrix.shroomrun.Scenario;
-import com.sulfrix.shroomrun.entities.Camera;
+import com.sulfrix.sulfur.Scenario;
+import com.sulfrix.sulfur.entity.Camera;
 import com.sulfrix.shroomrun.entities.Tile;
 import processing.core.PVector;
