+++ src/com/sulfrix/shroomrun/scenarios/TestScenario.java
@@ -1,14 +1,11 @@
 package com.sulfrix.shroomrun.scenarios;
 
-import com.sulfrix.shroomrun.Scenario;
-import com.sulfrix.shroomrun.entities.Camera;
+import com.sulfrix.sulfur.Scenario;
+import com.sulfrix.sulfur.entity.Camera;
 import com.sulfrix.shroomrun.entities.PhysicsTest;
 import com.sulfrix.shroomrun.entities.Tile;
-import com.sulfrix.shroomrun.lib.BoundingBox;
 import processing.core.PGraphics;
 import processing.core.PVector;
-
-import java.awt.*;
 
 public class TestScenario extends Scenario {
