+++ src/com/sulfrix/sulfur/Scenario.java
@@ -1,7 +1,7 @@
-package com.sulfrix.shroomrun;
+package com.sulfrix.sulfur;
 
 
-import com.sulfrix.shroomrun.lib.input.Input;
+import com.sulfrix.sulfur.lib.input.Input;
 import processing.core.PApplet;
 import processing.core.PGraphics;
 
@@ -17,9 +17,10 @@ public abstract class Scenario {
 
     public boolean updatedOnce;
 
+    public boolean initialized;
+
     public Scenario() {
         world = new World();
-        init();
     }
 
     public abstract void init();
