+++ src/com/sulfrix/sulfur/lib/BoundingBox.java
@@ -1,4 +1,4 @@
-package com.sulfrix.shroomrun.lib;
+package com.sulfrix.sulfur.lib;
 
 import processing.core.PVector;
