+++ src/com/sulfrix/sulfur/lib/GlobalManagers/RNG.java
@@ -1,4 +1,4 @@
-package com.sulfrix.shroomrun.lib.GlobalManagers;
+package com.sulfrix.sulfur.lib.GlobalManagers;
 
 import processing.core.PApplet;
 
@@ -8,11 +8,6 @@ public abstract class RNG extends GlobalManager {
 
     public static int seed;
     public static int offset;
-
-    public static void init(PApplet ownerApplet) {
-        owner = ownerApplet;
-        Scramble();
-    }
 
     public static int RandomInt(int min, int max) {
         var rand = new Random();
