+++ src/com/sulfrix/sulfur/lib/RenderPosType.java
@@ -1,4 +1,4 @@
-package com.sulfrix.shroomrun.lib;
+package com.sulfrix.sulfur.lib;
 
 public enum RenderPosType {
     WORLD_SPACE,
